// backend/src/database/generated/lobby.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: Rd9u+w9i5U+76YP4YQUExwUtoF/pszt/0pe483Lxy/Z3UQhqZXZKpl0S2z6dtV+WvhIkdkjiV7YZrw1SZVae0Q==
 */

/* eslint-disable */
// tslint:disable

import GameOptions from './game_options'

interface Lobby {
  date_created: Date
  game: (number) | null
  game_name: string
  game_options: (GameOptions['id']) | null
  /**
   * @default nextval('lobby_id_seq'::regclass)
   */
  id: number & {readonly __brand?: 'lobby_id'}
  max_players: number
  min_players: number
  players: (Array<number | null>) | null
  state: string
}
export default Lobby;

interface Lobby_InsertParameters {
  date_created: Date
  game?: (number) | null
  game_name: string
  game_options?: (GameOptions['id']) | null
  /**
   * @default nextval('lobby_id_seq'::regclass)
   */
  id?: number & {readonly __brand?: 'lobby_id'}
  max_players: number
  min_players: number
  players?: (Array<number | null>) | null
  state: string
}
export type {Lobby_InsertParameters}